Remove inline update functions from useMutations
